Thumbnails

+++ app/Http/Controllers/CommunityPostController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StorePostRequest;
 use App\Models\Community;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Intervention\Image\Facades\Image;
 
 class CommunityPostController extends Controller
 {
@@ -49,6 +50,12 @@ class CommunityPostController extends Controller
             $request->file('post_image')
                 ->storeAs('posts/' . $post->id, $image);
             $post->update(['post_image' => $image]);
+
+            $file = Image::make(storage_path('app/public/posts/' . $post->id . '/' . $image));
+            $file->resize(600, null, function ($constraint) {
+                $constraint->aspectRatio();
+            });
+            $file->save(storage_path('app/public/posts/' . $post->id . '/thumbnail_' . $image));
         }
 
         return redirect()->route('communities.show', $community);
@@ -100,11 +107,17 @@ class CommunityPostController extends Controller
             $request->file('post_image')
                 ->storeAs('posts/' . $post->id, $image);
 
-            if ($post->post_image != '') {
+            if ($post->post_image != '' && $post->post_image != $image) {
                 unlink(storage_path('app/public/posts/' . $post->id . '/' . $post->post_image));
             }
 
             $post->update(['post_image' => $image]);
+
+            $file = Image::make(storage_path('app/public/posts/' . $post->id . '/' . $image));
+            $file->resize(600, null, function ($constraint) {
+                $constraint->aspectRatio();
+            });
+            $file->save(storage_path('app/public/posts/' . $post->id . '/thumbnail_' . $image));
         }
 
         return redirect()->route('communities.posts.show', [$community, $post]);
